Extract most of pubuspec check into a validator
